pyproject.toml
[project]
name = "DeepResearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.4.247",
    "firecrawl>=1.10.2",
    "google-genai>=0.6.0",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "markitdown>=0.0.1a3",
    "pydantic-ai[logfire]>=0.0.19",
    "pydantic-ai-slim[openai]>=0.0.19",
    "pydantic-settings>=2.7.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "logfire[httpx]>=3.4.0",
    "langchain-google-genai>=2.0.9",
    "ipywidgets>=8.1.5",
    "aiohttp>=3.11.12",
    "beautifulsoup4>=4.13.3",
]
